// src/main/app/src/app/zoeken/model/zoek-veld.ts
/*
 * SPDX-FileCopyrightText: 2021 Atos
 * SPDX-License-Identifier: EUPL-1.2+
 */

export enum ZoekVeld {
  ALLE = "ALLE",

  ZAAK_IDENTIFICATIE = "ZAAK_IDENTIFICATIE",
  ZAAK_OMSCHRIJVING = "ZAAK_OMSCHRIJVING",
  ZAAK_TOELICHTING = "ZAAK_TOELICHTING",
  ZAAK_INITIATOR = "ZAAK_INITIATOR",
  ZAAK_BETROKKENE_ADVISEUR = "ZAAK_BETROKKENE_ADVISEUR",
  ZAAK_BETROKKENE_BELANGHEBBENDE = "ZAAK_BETROKKENE_BELANGHEBBENDE",
  ZAAK_BETROKKENE_BESLISSER = "ZAAK_BETROKKENE_BESLISSER",
  ZAAK_BETROKKENE_KLANTCONTACTER = "ZAAK_BETROKKENE_KLANTCONTACTER",
  ZAAK_BETROKKENE_ZAAKCOORDINATOR = "ZAAK_BETROKKENE_ZAAKCOORDINATOR",
  ZAAK_BETROKKENE_MEDE_INITIATOR = "ZAAK_BETROKKENE_MEDE_INITIATOR",
  ZAAK_BETROKKENEN = "ZAAK_BETROKKENEN",
  ZAAK_BAGOBJECTEN = "ZAAK_BAGOBJECTEN",

  TAAK_ZAAK_ID = "TAAK_ZAAK_ID",
  TAAK_TOELICHTING = "TAAK_TOELICHTING",

  DOCUMENT_TITEL = "DOCUMENT_TITEL",
  DOCUMENT_BESCHRIJVING = "DOCUMENT_BESCHRIJVING",
}